requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
kombu==5.4.2
nested-multipart-parser==1.5.0
packaging==24.2
phonenumbers==8.13.54
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
redis-cli==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
